+++ Viajante/src/aseleccion/Aseleccion_Truncamiento.java
@@ -13,19 +13,19 @@ public class Aseleccion_Truncamiento implements ASeleccion {
 	//El algoritmo de selección por truncamiento elige el top % de una población y llena 
 	//la nueva población con esos individuos
 	
-	public void seleccionar(Cromosoma[] poblacion, float[] aptitudes, Cromosoma[] seleccionados) {
+	public void seleccionar(Cromosoma[] poblacion, int[] aptitudes, Cromosoma[] seleccionados) {
 		
 		int aCopiar = (int)(poblacion.length * _porcentaje);
 		int numCopias = poblacion.length / aCopiar;
-		Comparator<javafx.util.Pair<Float, Integer>> comparador;
+		Comparator<javafx.util.Pair<Integer, Integer>> comparador;
 		//Para arreglar el fallo de JavaFX:
 		//JRE System Libraries -> Click Derecho -> Propiedades, Cambiar a "Workspace Default" (3era opción)
-		comparador = new Comparator<javafx.util.Pair<Float, Integer>>()
+		comparador = new Comparator<javafx.util.Pair<Integer, Integer>>()
 		{
 
 			@Override
-			public int compare(Pair<Float, Integer> o1,
-					Pair<Float, Integer> o2) {
+			public int compare(Pair<Integer, Integer> o1,
+					Pair<Integer, Integer> o2) {
 				if(o1.getKey() > o2.getKey()) return -1;
 				if(o1.getKey() == o2.getKey()) return 0;
 				if(o1.getKey() < o2.getKey()) return 1;
@@ -33,10 +33,10 @@ public class Aseleccion_Truncamiento implements ASeleccion {
 			}
 		};
 		
-		PriorityQueue<javafx.util.Pair<Float, Integer>> monticuloMaximos = new PriorityQueue<javafx.util.Pair<Float, Integer>>(comparador);
+		PriorityQueue<javafx.util.Pair<Integer, Integer>> monticuloMaximos = new PriorityQueue<javafx.util.Pair<Integer, Integer>>(comparador);
 		for(int i = 0; i < aptitudes.length; i++)
 		{
-			monticuloMaximos.add(new javafx.util.Pair<Float, Integer>(aptitudes[i], i));
+			monticuloMaximos.add(new javafx.util.Pair<Integer, Integer>(aptitudes[i], i));
 		}
 		
 		//Obtener indices ordenados
